"Se activa consulta filtrando los hoteles por el país argentina"
